requirements_pose.txt
absl-py==2.0.0
asttokens==2.4.1
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.7
cython-bbox==0.1.3
decorator==5.1.1
easydict==1.11
exceptiongroup==1.2.0
executing==2.0.1
faiss==1.7.4
ffmpeg==1.4
filelock==3.13.1
fonttools==4.47.0
freetype-py==2.4.0
fsspec==2023.12.2
gdown==4.7.1
google-auth==2.25.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
hsluv==5.0.4
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
lap==0.4.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prettytable==3.9.0
prompt-toolkit==3.0.43
protobuf==4.25.1
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt5Designer==5.14.1
pyrealsense2==2.54.2.5684
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smplx==0.1.28
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.9.12
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vispy==0.14.1
wcwidth==0.2.12
Werkzeug==3.0.1
yacs==0.1.8
zipp==3.17.0
